+++ src/main/java/com/github/jlangch/venice/impl/docgen/cheatsheet/section/IoFileSection.java
@@ -44,6 +44,7 @@ public class IoFileSection implements ISectionBuilder {
         file.addItem(diBuilder.getDocItem("io/file"));
         file.addItem(diBuilder.getDocItem("io/file-parent"));
         file.addItem(diBuilder.getDocItem("io/file-name"));
+        file.addItem(diBuilder.getDocItem("io/file-basename"));
         file.addItem(diBuilder.getDocItem("io/file-path"));
         file.addItem(diBuilder.getDocItem("io/file-absolute"));
         file.addItem(diBuilder.getDocItem("io/file-canonical"));